Limpiando register y login y eliminando
